==> 2. Membuat-valid218.ipynb <==
!git clone https://github.com/FuadHamdiBahar/EKGCroppedData.git

import pandas as pd

CATEGORICAL_COLUMNS = ['AF', 'HRclass', 'Regularity', 'Qpath', 'Qpathlocation', 'STsegment',
                       'STTchanges', 'STTiskemia', 'STTnonspesific', 'STloc', 'Twave', 'Twaveloc',
                       'VES', 'AES', 'JES', 'Uwave']
PREDICTOR_COLUMNS = ['ID','AF','HR','HRclass','Regularity','Qpath','Qpathlocation',
                     'STsegment', 'STTchanges','STTiskemia','STTnonspesific','STloc',
                     'QTinterval', 'QTintervalcorrected', 'Twave', 'Twaveloc', 'Taxisdegree',
                     'VES', 'AES', 'JES', 'Uwave']

df = pd.read_csv('./EKGCroppedData/valid128.csv')
df = df.drop(columns = ['Unnamed: 0'])
df = df[PREDICTOR_COLUMNS]
# df.info()

for i in CATEGORICAL_COLUMNS:
    print(i, set(df[i].to_list()), '\t\t\t', len(set(df[i].to_list())))

def cek(dataframe):
    for col in CATEGORICAL_COLUMNS:
        print(dataframe[col].value_counts())

from imblearn.over_sampling import SMOTE, RandomOverSampler
from imblearn.under_sampling import RandomUnderSampler


s = SMOTE()
ros = RandomOverSampler(sampling_strategy = {1:100, 4:100})
df, y = ros.fit_resample(df, df['Twaveloc'])

cek(df)

for col in CATEGORICAL_COLUMNS:
    for label in df[col].unique():
        query = df[df[col] == label].index
        if len(query) < 100:
            print(col, label, len(query))

valid_index = []
for col in CATEGORICAL_COLUMNS:
    for label in df[col].unique():
        query = df[col][df[col] == label].index
        
        if len(query) < 700:
#             print(col, label, len(query))
            valid_index.append(query.to_list())

from itertools import chain
import numpy as np

valid_index = list(chain.from_iterable(valid_index))
valid_index = np.unique(valid_index)
# valid_index

valid_data = df.loc[valid_index, :].reset_index(drop=True)

valid_data.head()

valid_data.info()

valid_data.to_csv('valid218.csv')
