==> 3. Membuat valid248.ipynb <==
import pandas as pd

# DEFINISI VARIABEL UMUM
RANDOM_STATE = 42

# sementara hapus Brugada
predictor_col =  ['AF','HR','HRclass','Regularity','Qpath','Qpathlocation', 
                  'STsegment', 'STTchanges','STTiskemia','STTnonspesific','STloc',
                  'QTinterval', 'QTintervalcorrected', 'Twave', 'Twaveloc', 'Taxisdegree',
                  'VES', 'AES', 'JES', 'Uwave']

# sementara hapus RVH WPW AVBLOCK SABLOCK AFLUT SVT
response_col = ['Sinus', 'Tachy', 'Brady', 'LVH', 'QRSfragmented', 'Lowvoltage', 
                'STEMI','BER', 'Taxis', 'Normal_ECG']

df = pd.read_csv("EKGCroppedData/CSV_PING.csv")
df = df.drop(columns = ['Unnamed: 0','Unnamed: 0.1','use', 'ID', 'Gender', 'age_55'])

data = df.copy()

valid_index = []

for col in response_col:
    for label in data[col].unique():
        query = data[col][data[col] == label].index
        
#         if len(query) < 1500:
#             valid_index.append(query.to_list())
        print(col, label, len(query))

from itertools import chain
import numpy as np

valid_index = list(chain.from_iterable(valid_index))
valid_index = np.unique(valid_index)
# valid_index

valid_data = data.loc[valid_index, :].reset_index(drop=True)

valid_data.to_csv('valid248.csv')
