==> 4. Membuat valid248 Resample.ipynb <==
import pandas as pd

df = pd.read_csv('valid248.csv').drop(columns = 'Unnamed: 0')

df[all_col].info()

# DEFINISI VARIABEL UMUM
RANDOM_STATE = 42

ALL_COL = [
    'AF','HR','HRclass','Regularity','Qpath','Qpathlocation', 'STsegment', 'STTchanges','STTiskemia','STTnonspesific','STloc',
    'QTinterval', 'QTintervalcorrected', 'Twave', 'Twaveloc', 'Taxisdegree', 'VES', 'AES', 'JES', 'Uwave', 'Sinus', 'Tachy', 
    'Brady', 'LVH', 'QRSfragmented', 'Lowvoltage', 'STEMI','BER', 'Taxis', 'Normal_ECG'
]

# sementara hapus Brugada
PREDICTOR_COL =  ['AF','HR','HRclass','Regularity','Qpath','Qpathlocation', 
                  'STsegment', 'STTchanges','STTiskemia','STTnonspesific','STloc',
                  'QTinterval', 'QTintervalcorrected', 'Twave', 'Twaveloc', 'Taxisdegree',
                  'VES', 'AES', 'JES', 'Uwave']

# sementara hapus RVH WPW AVBLOCK SABLOCK AFLUT SVT
RESPONSE_COL = ['Sinus', 'Tachy', 'Brady', 'LVH', 'QRSfragmented', 'Lowvoltage', 
                'STEMI','BER', 'Taxis', 'Normal_ECG']

data = df.copy()
data = data[all_col]

for i in RESPONSE_COL:
    print(i, set(data[i].to_list()), '\t\t\t', len(set(data[i].to_list())))

for col in RESPONSE_COL:
    print(data[col].value_counts())

from imblearn.over_sampling import SMOTE, RandomOverSampler
from imblearn.under_sampling import RandomUnderSampler

s = SMOTE()
ros = RandomOverSampler(sampling_strategy = {1:100, 2:100, 3:100})
data, y = ros.fit_resample(data, data['Taxis'])

data.to_csv('valid248resample.csv')

data.info()
