==> 5. CNN Model.ipynb <==
!git clone https://github.com/FuadHamdiBahar/EKGCroppedData.git

import tensorflow as tf
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

IMG_RES = 32
RANDOM_STATE = 42

CATEGORICAL_COLUMNS = [
    'AF', 'HRclass', 'Regularity', 'Qpath', 'Qpathlocation', 'STsegment', 'STTchanges',
    'STTiskemia', 'STTnonspesific', 'STloc', 'Twave', 'Twaveloc','VES', 'AES', 'JES', 'Uwave'
]

df = pd.read_csv('EKGCroppedData/valid218.csv').drop(columns = 'Unnamed: 0')
df['STTiskemia'].replace({1:0, 2:1}, inplace = True)
df['STTnonspesific'].replace({1:0, 2:1}, inplace = True)

df.info()

from itertools import chain

test_index = []
for col in CATEGORICAL_COLUMNS:
    for label in df[col].unique():
        query = df[col][df[col] == label]
        sample_index = query.sample(n = 2, random_state = RANDOM_STATE, replace = False).index.to_list()
        test_index.append(sample_index)

flatten_test_index = list(chain.from_iterable(test_index))
unique_flatten_test_index = np.unique(flatten_test_index)

train_index = [i for i in range(len(df)) if i not in unique_flatten_test_index]

df_train = df.loc[train_index, :].reset_index(drop=True)
df_test = df.loc[unique_flatten_test_index, :].reset_index(drop=True)

X_train = df_train.loc[:, 'ID']
X_test = df_test.loc[:, 'ID']

from tqdm import tqdm

def preprocessing(folder, data):
    list_image_arr = []
    for nama in tqdm(data):
        path = 'EKGCroppedData/CROPPED_12_PING_IMG/{}/{}.jpeg'.format(folder, nama)
        image = tf.keras.preprocessing.image.load_img(path)
        image_arr = tf.keras.preprocessing.image.img_to_array(image)
        resize = tf.image.resize(image_arr, [IMG_RES, IMG_RES])
        list_image_arr.append(resize)
    return np.array(list_image_arr)

aVF_train = preprocessing('aVF', X_train)
aVL_train = preprocessing('aVL', X_train)
aVR_train = preprocessing('aVR', X_train)
I_train = preprocessing('I', X_train)
II_train = preprocessing('II', X_train)
III_train = preprocessing('III', X_train)
V1_train = preprocessing('V1', X_train)
V2_train = preprocessing('V2', X_train)
V3_train = preprocessing('V3', X_train)
V4_train = preprocessing('V4', X_train)
V5_train = preprocessing('V5', X_train)
V6_train = preprocessing('V6', X_train)

aVF_test = preprocessing('aVF', X_test)
aVL_test = preprocessing('aVL', X_test)
aVR_test = preprocessing('aVR', X_test)
I_test = preprocessing('I', X_test)
II_test = preprocessing('II', X_test)
III_test = preprocessing('III', X_test)
V1_test = preprocessing('V1', X_test)
V2_test = preprocessing('V2', X_test)
V3_test = preprocessing('V3', X_test)
V4_test = preprocessing('V4', X_test)
V5_test = preprocessing('V5', X_test)
V6_test = preprocessing('V6', X_test)

aVF_train = aVF_train / 255.0 
aVL_train = aVL_train / 255.0 
aVR_train = aVR_train / 255.0 
I_train = I_train / 255.0 
II_train = II_train / 255.0 
III_train = III_train / 255.0 
V1_train = V1_train / 255.0 
V2_train = V2_train / 255.0 
V3_train = V3_train / 255.0 
V4_train = V4_train / 255.0 
V5_train = V5_train / 255.0 
V6_train = V6_train / 255.0 

aVF_test = aVF_test / 255.0 
aVL_test = aVL_test / 255.0 
aVR_test = aVR_test / 255.0 
I_test = I_test / 255.0 
II_test = II_test / 255.0 
III_test = III_test / 255.0 
V1_test = V1_test / 255.0 
V2_test = V2_test / 255.0 
V3_test = V3_test / 255.0 
V4_test = V4_test / 255.0 
V5_test = V5_test / 255.0 
V6_test = V6_test / 255.0 

# w = 10
# h = 10
# fig = plt.figure(figsize=(12, 12))
# columns = 4
# rows = 3

# fig.add_subplot(rows, columns, 1)
# plt.imshow(I_train[0])
# fig.add_subplot(rows, columns, 2)
# plt.imshow(aVR_train[0])
# fig.add_subplot(rows, columns, 3)
# plt.imshow(V1_train[0])
# fig.add_subplot(rows, columns, 4)
# plt.imshow(V4_train[0])

# fig.add_subplot(rows, columns, 5)
# plt.imshow(II_train[0])
# fig.add_subplot(rows, columns, 6)
# plt.imshow(aVL_train[0])
# fig.add_subplot(rows, columns, 7)
# plt.imshow(V2_train[0])
# fig.add_subplot(rows, columns, 8)
# plt.imshow(V5_train[0])

# fig.add_subplot(rows, columns, 9)
# plt.imshow(III_train[0])
# fig.add_subplot(rows, columns, 10)
# plt.imshow(aVF_train[0])
# fig.add_subplot(rows, columns, 11)
# plt.imshow(V3_train[0])
# fig.add_subplot(rows, columns, 12)
# plt.imshow(V6_train[0])

af_train = df_train.loc[:, 'AF']
hr_train = df_train.loc[:, 'HR']
hc_train = df_train.loc[:, 'HRclass']
re_train = df_train.loc[:, 'Regularity']
qp_train = df_train.loc[:, 'Qpath']
ql_train = df_train.loc[:, 'Qpathlocation']
ss_train = df_train.loc[:, 'STsegment']
sc_train = df_train.loc[:, 'STTchanges']
si_train = df_train.loc[:, 'STTiskemia']
sn_train = df_train.loc[:, 'STTnonspesific']
sl_train = df_train.loc[:, 'STloc']
qi_train = df_train.loc[:, 'QTinterval']
qc_train = df_train.loc[:, 'QTintervalcorrected']
tw_train = df_train.loc[:, 'Twave']
tl_train = df_train.loc[:, 'Twaveloc']
td_train = df_train.loc[:, 'Taxisdegree']
vs_train = df_train.loc[:, 'VES']
as_train = df_train.loc[:, 'AES']
js_train = df_train.loc[:, 'JES']
uw_train = df_train.loc[:, 'Uwave']

af_test = df_test.loc[:, 'AF']
hr_test = df_test.loc[:, 'HR']
hc_test = df_test.loc[:, 'HRclass']
re_test = df_test.loc[:, 'Regularity']
qp_test = df_test.loc[:, 'Qpath']
ql_test = df_test.loc[:, 'Qpathlocation']
ss_test = df_test.loc[:, 'STsegment']
sc_test = df_test.loc[:, 'STTchanges']
si_test = df_test.loc[:, 'STTiskemia']
sn_test = df_test.loc[:, 'STTnonspesific']
sl_test = df_test.loc[:, 'STloc']
qi_test = df_test.loc[:, 'QTinterval']
qc_test = df_test.loc[:, 'QTintervalcorrected']
tw_test = df_test.loc[:, 'Twave']
tl_test = df_test.loc[:, 'Twaveloc']
td_test = df_test.loc[:, 'Taxisdegree']
vs_test = df_test.loc[:, 'VES']
as_test = df_test.loc[:, 'AES']
js_test = df_test.loc[:, 'JES']
uw_test = df_test.loc[:, 'Uwave']

from tensorflow.keras.utils import to_categorical

af_train = to_categorical(af_train)
hr_train = hr_train.to_numpy()
hc_train = to_categorical(hc_train)
re_train = to_categorical(re_train)
qp_train = to_categorical(qp_train)
ql_train = to_categorical(ql_train)
ss_train = to_categorical(ss_train)
sc_train = to_categorical(sc_train)
si_train = to_categorical(si_train)
sn_train = to_categorical(sn_train)
sl_train = to_categorical(sl_train)
qi_train = qi_train.to_numpy()
qc_train = qc_train.to_numpy()
tw_train = to_categorical(tw_train)
tl_train = to_categorical(tl_train)
td_train = td_train.to_numpy()
vs_train = to_categorical(vs_train)
as_train = to_categorical(as_train)
js_train = to_categorical(js_train)
uw_train = to_categorical(uw_train)

af_test = to_categorical(af_test)
hr_test = hr_test.to_numpy()
hc_test = to_categorical(hc_test)
re_test = to_categorical(re_test)
qp_test = to_categorical(qp_test)
ql_test = to_categorical(ql_test)
ss_test = to_categorical(ss_test)
sc_test = to_categorical(sc_test)
si_test = to_categorical(si_test)
sn_test = to_categorical(sn_test)
sl_test = to_categorical(sl_test)
qi_test = qi_test.to_numpy()
qc_test = qc_test.to_numpy()
tw_test = to_categorical(tw_test)
tl_test = to_categorical(tl_test)
td_test = td_test.to_numpy()
vs_test = to_categorical(vs_test)
as_test = to_categorical(as_test)
js_test = to_categorical(js_test)
uw_test = to_categorical(uw_test)

train_shape = [
    af_train.shape, hr_train.shape, hc_train.shape, re_train.shape, qp_train.shape, ql_train.shape, ss_train.shape,
    sc_train.shape, si_train.shape, sn_train.shape, sl_train.shape, qi_train.shape, qc_train.shape, tw_train.shape,
    tl_train.shape, td_train.shape, vs_train.shape, as_train.shape, js_train.shape, uw_train.shape
]
train_shape

test_shape = [
    af_test.shape, hr_test.shape, hc_test.shape, re_test.shape, qp_test.shape, ql_test.shape, ss_test.shape,
    sc_test.shape, si_test.shape, sn_test.shape, sl_test.shape, qi_test.shape, qc_test.shape, tw_test.shape,
    tl_test.shape, td_test.shape, vs_test.shape, as_test.shape, js_test.shape, uw_test.shape
]
test_shape

def mobilenet():
    b_model = tf.keras.applications.MobileNetV2(weights = 'imagenet', include_top = False)
    b_model.trainable = False
    x = tf.keras.layers.GlobalAveragePooling2D()(b_model.output)
    out = tf.keras.layers.Dense(1) (x)
    
    model = tf.keras.Model(b_model.input, out)
    model.compile(optimizer = 'adam', loss = 'mse')
    
    return model

main_model = mobilenet()

pretrained_model = tf.keras.Model(main_model.input, main_model.layers[-3].output)
# pretrained_model.summary()

I_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'I_inp')
II_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'II_inp')
III_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'III_inp')
aVR_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'aVR_inp')
aVL_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'aVL_inp')
aVF_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'aVF_inp')
V1_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'V1_inp')
V2_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'V2_inp')
V3_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'V3_inp')
V4_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'V4_inp')
V5_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'V5_inp')
V6_inp = tf.keras.Input(shape = (IMG_RES, IMG_RES, 3), name = 'V6_inp')

I_model = pretrained_model(I_inp) 
II_model = pretrained_model(II_inp) 
III_model = pretrained_model(III_inp) 
aVR_model = pretrained_model(aVR_inp) 
aVL_model = pretrained_model(aVL_inp) 
aVF_model = pretrained_model(aVF_inp) 
V1_model = pretrained_model(V1_inp) 
V2_model = pretrained_model(V2_inp) 
V3_model = pretrained_model(V3_inp) 
V4_model = pretrained_model(V4_inp) 
V5_model = pretrained_model(V5_inp) 
V6_model = pretrained_model(V6_inp)

concat_model = [
    I_model, II_model, III_model, aVR_model, aVL_model, aVF_model,
    V1_model, V2_model, V3_model, V4_model, V5_model, V6_model
]

concat = tf.keras.layers.concatenate(concat_model)
flatten = tf.keras.layers.Flatten() (concat)

# AF
af_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
af_dense = tf.keras.layers.Dropout(0.5) (af_dense)
af_out = tf.keras.layers.Dense(2, activation = 'softmax', name='AF') (af_dense)

# HR
hr_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
hr_dense = tf.keras.layers.Dropout(0.5) (hr_dense)
hr_out = tf.keras.layers.Dense(1, activation = 'linear', name='HR') (hr_dense)

# HRclass
hc_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
hc_dense = tf.keras.layers.Dropout(0.5) (hc_dense)
hc_out = tf.keras.layers.Dense(3, activation = 'softmax', name='HRclass') (hc_dense)

# Regularity
re_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
re_dense = tf.keras.layers.Dropout(0.5) (re_dense)
re_out = tf.keras.layers.Dense(2, activation = 'softmax', name='Regularity') (re_dense)

# Qpath
qp_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
qp_dense = tf.keras.layers.Dropout(0.5) (qp_dense)
qp_out = tf.keras.layers.Dense(2, activation = 'softmax', name='Qpath') (qp_dense)

# Qpathlocation
ql_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
ql_dense = tf.keras.layers.Dropout(0.5) (ql_dense)
ql_out = tf.keras.layers.Dense(3, activation = 'softmax', name='Qpathlocation') (ql_dense)

# STsegment
ss_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
ss_dense = tf.keras.layers.Dropout(0.5) (ss_dense)
ss_out = tf.keras.layers.Dense(4, activation = 'softmax', name='STsegment') (ss_dense)

# STTchanges
sc_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
sc_dense = tf.keras.layers.Dropout(0.5) (sc_dense)
sc_out = tf.keras.layers.Dense(3, activation = 'softmax', name='STTchanges') (sc_dense)

# STTiskemia
si_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
si_dense = tf.keras.layers.Dropout(0.5) (si_dense)
si_out = tf.keras.layers.Dense(2, activation = 'softmax', name='STTiskemia') (si_dense)

# STTnonspesific
sn_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
sn_dense = tf.keras.layers.Dropout(0.5) (sn_dense)
sn_out = tf.keras.layers.Dense(2, activation = 'softmax', name='STTnonspesific') (sn_dense)

# STloc
sl_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
sl_dense = tf.keras.layers.Dropout(0.5) (sl_dense)
sl_out = tf.keras.layers.Dense(5, activation = 'softmax', name='STloc') (sl_dense)

# QTinterval
qi_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
qi_dense = tf.keras.layers.Dropout(0.5) (qi_dense)
qi_out = tf.keras.layers.Dense(1, activation = 'linear', name='QTinterval') (qi_dense)

# QTintervalcorrected
qc_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
qc_dense = tf.keras.layers.Dropout(0.5) (qc_dense)
qc_out = tf.keras.layers.Dense(1, activation = 'linear', name='QTintervalcorrected') (qc_dense)

# Twave
tw_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
tw_dense = tf.keras.layers.Dropout(0.5) (tw_dense)
tw_out = tf.keras.layers.Dense(4, activation = 'softmax', name='Twave') (tw_dense)

# Twaveloc
tl_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
tl_dense = tf.keras.layers.Dropout(0.5) (tl_dense)
tl_out = tf.keras.layers.Dense(5, activation = 'softmax', name='Twaveloc') (tl_dense)

# Taxisdegree
td_dense = tf.keras.layers.Dense(128, activation = 'relu', kernel_initializer='he_uniform') (flatten)
td_dense = tf.keras.layers.Dropout(0.5) (td_dense)
td_out = tf.keras.layers.Dense(1, activation = 'linear', name='Taxisdegree') (td_dense)

# VES
vs_dense = tf.keras.layers.Dense(
    128, activation='relu', kernel_initializer='he_uniform')(flatten)
vs_dense = tf.keras.layers.Dropout(0.5)(vs_dense)
vs_out = tf.keras.layers.Dense(2, activation='softmax', name='VES')(vs_dense)

# AES
as_dense = tf.keras.layers.Dense(
    128, activation='relu', kernel_initializer='he_uniform')(flatten)
as_dense = tf.keras.layers.Dropout(0.5)(as_dense)
as_out = tf.keras.layers.Dense(2, activation='softmax', name='AES')(as_dense)

# JES
js_dense = tf.keras.layers.Dense(
    128, activation='relu', kernel_initializer='he_uniform')(flatten)
js_dense = tf.keras.layers.Dropout(0.5)(js_dense)
js_out = tf.keras.layers.Dense(2, activation='softmax', name='JES')(js_dense)

# Uwave
uw_dense = tf.keras.layers.Dense(
    128, activation='relu', kernel_initializer='he_uniform')(flatten)
uw_dense = tf.keras.layers.Dropout(0.5)(uw_dense)
uw_out = tf.keras.layers.Dense(2, activation='softmax', name='Uwave')(uw_dense)

all_inputs = [I_inp, II_inp, III_inp, aVR_inp, aVL_inp, aVF_inp, V1_inp, V2_inp, V3_inp, V4_inp, V5_inp, V6_inp]
all_outputs = [af_out, hr_out, hc_out, re_out, qp_out, ql_out, ss_out, sc_out, si_out, sn_out,
               sl_out, qi_out, qc_out, tw_out, tl_out, td_out, vs_out, as_out, js_out, uw_out]

losses = {
    'AF': 'categorical_crossentropy',
    'HR': 'mse',
    'HRclass': 'categorical_crossentropy',
    'Regularity': 'categorical_crossentropy',
    'Qpath': 'categorical_crossentropy',
    'Qpathlocation': 'categorical_crossentropy',
    'STsegment': 'categorical_crossentropy',
    'STTchanges': 'categorical_crossentropy',
    'STTiskemia': 'categorical_crossentropy',
    'STTnonspesific': 'categorical_crossentropy',
    'STloc': 'categorical_crossentropy',
    'QTinterval': 'mse',
    'QTintervalcorrected': 'mse',
    'Twave': 'categorical_crossentropy',
    'Twaveloc': 'categorical_crossentropy',
    'Taxisdegree': 'mse',
    'VES': 'categorical_crossentropy',
    'AES': 'categorical_crossentropy',
    'JES': 'categorical_crossentropy',
    'Uwave': 'categorical_crossentropy'
}

all_metrics = {
    'AF': 'accuracy',
    'HR': 'mse',
    'HRclass': 'accuracy',
    'Regularity': 'accuracy',
    'Qpath': 'accuracy',
    'Qpathlocation': 'accuracy',
    'STsegment': 'accuracy',
    'STTchanges': 'accuracy',
    'STTiskemia': 'accuracy',
    'STTnonspesific': 'accuracy',
    'STloc': 'accuracy',
    'QTinterval': 'mse',
    'QTintervalcorrected': 'mse',
    'Twave': 'accuracy',
    'Twaveloc': 'accuracy',
    'Taxisdegree': 'mse',
    'VES': 'accuracy',
    'AES': 'accuracy',
    'JES': 'accuracy',
    'Uwave': 'accuracy'
}

base_model = tf.keras.Model(inputs = all_inputs, outputs = all_outputs)
base_model.compile(loss=losses, optimizer='adam', metrics=all_metrics)
# base_model.summary()

tf.keras.utils.plot_model(base_model)

all_x_train = [
    I_train, II_train, III_train, aVR_train, aVL_train, aVF_train, 
    V1_train, V2_train, V3_train, V4_train, V5_train, V6_train
]
all_x_test = [
    I_test, II_test, III_test, aVR_test, aVL_test, aVF_test,
    V1_test, V2_test, V3_test, V4_test, V5_test, V6_test
]

all_y_train = [
    af_train, hr_train, hc_train, re_train, qp_train, ql_train, ss_train, sc_train, si_train, sn_train,
    sl_train, qi_train, qc_train, tw_train, tl_train, td_train, vs_train, as_train, js_train, uw_train
]

all_y_test = [
    af_test, hr_test, hc_test, re_test, qp_test, ql_test, ss_test, sc_test, si_test, sn_test,
    sl_test, qi_test, qc_test, tw_test, tl_test, td_test, vs_test, as_test, js_test, uw_test
]

h = base_model.fit(
    x = all_x_train,
    y = all_y_train,
    batch_size = 20,
    validation_data = (all_x_test, all_y_test),
    epochs = 10,
    verbose = 0
)

plt.plot(h.history['loss'])
plt.plot(h.history['val_loss'])

categorical_name = [
    'AF', 'HRclass', 'Regularity', 'Qpath', 'Qpathlocation', 'STsegment', 'STTchanges',
    'STTiskemia', 'STTnonspesific', 'STloc', 'Twave', 'Twaveloc', 'VES', 'AES', 'JES', 'Uwave'
]

from math import ceil
n_column = 4
n_row = ceil(len(categorical_name) / n_column)

fig = plt.figure(figsize=(20, 16))
axs = fig.subplots(nrows=n_row, ncols=n_column)

i = 0
for j in range(n_row):
    for k in range(n_column):
        try:
            if i == len(categorical_name):
                break
            axs[j, k].plot(h.history[categorical_name[i] + '_accuracy'], label = 'train')
            axs[j, k].plot(h.history['val_' + categorical_name[i] + '_accuracy'], label = 'test')
            axs[j, k].set_title(categorical_name[i])
            axs[j, k].legend(loc ="lower right")
        except KeyError as e:
            pass
        i += 1

linear_name = [
    'HR', 'QTinterval', 'QTintervalcorrected', 'Taxisdegree'
]

from math import ceil
n_column = 4
n_row = 1

fig = plt.figure(figsize=(20, 4))
axs = fig.subplots(nrows=n_row, ncols=n_column)

i = 0
for k in range(n_column):
    try:
        if i == len(linear_name):
            break
        axs[k].plot(h.history[linear_name[i] + '_mse'], label = 'train')
        axs[k].plot(h.history['val_' + linear_name[i] + '_mse'], label = 'test')
        axs[k].set_title(linear_name[i])
        axs[k].legend(loc ="lower right")
    except KeyError as e:
        pass
    i += 1

from sklearn.metrics import confusion_matrix, precision_score, recall_score, f1_score
import seaborn as sns

def categorical_evaluation(actual, pred):
    pred = np.argmax(pred, axis = 1)
    actual = np.argmax(actual, axis = 1)
    
    p = precision_score(actual, pred, average='macro')
    r = recall_score(actual, pred, average='macro')
    f = f1_score(actual, pred, average='macro')

    print("Precision: {}, Recall: {}, F1: {}".format(p, r ,f))

    cm = confusion_matrix(actual, pred)
    sns.heatmap(cm, annot=True, cmap='Blues')

from sklearn.metrics import mean_squared_error, mean_absolute_error
from math import sqrt

def linear_evaluation(actual, pred):
    mse = mean_squared_error(actual, pred.flatten())
    rmse = sqrt(mse)
    mae = mean_absolute_error(actual, pred.flatten())
    print("MSE: {}, RMSE: {}, MAE: {}".format(mse, rmse, mae))

pred = base_model.predict(all_x_test)

categorical_evaluation(af_test, pred[0])

linear_evaluation(hr_test, pred[1])

categorical_evaluation(hc_test, pred[2])

categorical_evaluation(re_test, pred[3])

categorical_evaluation(qp_test, pred[4])

categorical_evaluation(ql_test, pred[5])

categorical_evaluation(ss_test, pred[6])

categorical_evaluation(sc_test, pred[7])

categorical_evaluation(si_test, pred[8])

categorical_evaluation(sn_test, pred[9])

categorical_evaluation(sl_test, pred[10])

linear_evaluation(qi_test, pred[11])

linear_evaluation(qc_test, pred[12])

categorical_evaluation(tw_test, pred[13])

categorical_evaluation(tl_test, pred[14])

linear_evaluation(td_test, pred[15])

categorical_evaluation(vs_test, pred[16])

categorical_evaluation(as_test, pred[17])

categorical_evaluation(js_test, pred[18])

categorical_evaluation(uw_test, pred[19])

print("GOOD JOB FUAD! ALHAMDULILLAH")
